fix: no longer necessary to overselect past end of final visual line
If selecting text with visual mode ('v') to complete, then it was necessary to
select past the end of the final line, otherwise the completion could be
appended just before the final character of the selection. This commit fixes
makes it so that is no longer required and also removes some off-by-one
nonsense from the code.